Enable parentheses-less function argument syntax
I am not sure about this. This is a unique "feature" for commas within an argument list: the list of iterators in the generator take priority over any subsequent arguments. The resulting syntax is very nice for the one dimensional case: sum(x for x=1:100), but very jarring for multiple dimensions: sum(x*y for x=1:100, y=1:100) -- is y a keyword argument? It would probably be best to require the extra parentheses to make the comma apply to the generator.