redefine TopNode so that there can be multiple independent top modules
making inference separate from base required modifying the definition of
TopNode to be more general, since Base may not even exist at points and
it is desirable for Core.Inference to be wholly separate from Base.
rather than special-casing Base, this introduces the concept that any
module can indicate that it is valid as a top module and will define
appropriate functions, as expected by the julia parser, to make this
legal. At the start of the module definition,
modules can ccall jl_set_istopmod(isprimary) to indicate this.
The primary top module will be used to resolve the top module for any
code that is not contained in another top module.
Only one module can be the primary top module at a time,
typically it will be Base.